feat-map: fix register
